classification with tensorflow model (very rough straight forwarward implementation)

+++ BatInspector/ViewModel/ViewModel.cs
@@ -35,6 +35,10 @@ namespace BatInspector
 
   public class ViewModel
   {
+    public const int OPT_INSPECT = 1;
+    public const int OPT_CUT = 2;
+    public const int OPT_PREPARE = 4;
+    public const int OPT_PREDICT = 8;
 
     string _selectedDir;
     Project _prj;
@@ -87,7 +91,11 @@ namespace BatInspector
       _wav = new WavFile();
     }
 
-
+    public void updateReport()
+    {
+      if (File.Exists(_selectedDir + "report.csv"))
+        _analysis.read(_selectedDir + "report.csv");
+    }
 
     public void initProject(DirectoryInfo dir)
     {
@@ -207,28 +215,59 @@ namespace BatInspector
     /// <summary>
     /// start evaluation of bat species
     /// </summary>
-    public void startEvaluation()
+    public int startEvaluation(int options)
     {
+      int retVal = 2;
       if(Prj != null)
       {
-        string exe = "D:/bin/R-4.1.0/bin/Rscript.exe";
-        string wrkDir = "D:/prj/bioacoustics";
-        string reportName = _selectedDir + "/report.csv";
-        string args = "cm.R " + _selectedDir + "/Records " + reportName;
-        DebugLog.log("starting evaluation: " + exe + " " + args, enLogType.INFO);
-        if (File.Exists(reportName))
+        
+        string exeR = "\"C:/Program Files/R/R-4.2.0/bin/Rscript.exe\"";
+        string reportName = _selectedDir + "report.csv";
+        reportName = reportName.Replace("\\", "/");
+        string specFile = "C:/Users/chrmu/bat/tierSta/species.csv";
+
+        if ((options & OPT_INSPECT) != 0)
         {
-          DebugLog.log("backup file: " + reportName, enLogType.INFO);
-          try
+          string scriptR = "C:/Users/chrmu/prj/BatInspector/R/features.R";
+          string argsR = scriptR + " " + _selectedDir + "/Records " + reportName + " " + specFile + " 312500";
+          DebugLog.log("starting evaluation of calls: " + exeR + " " + argsR, enLogType.INFO);
+          if (File.Exists(reportName))
           {
-            File.Delete(reportName + "_old");
-            File.Copy(reportName, reportName + "_old");
-            File.Delete(reportName);
+            DebugLog.log("backup file: " + reportName, enLogType.INFO);
+            try
+            {
+              File.Delete(reportName + "_old");
+              File.Copy(reportName, reportName + "_old");
+              File.Delete(reportName);
+            }
+            catch { }
           }
-          catch { }
+          retVal = _proc.LaunchCommandLineApp(exeR, null, _selectedDir, true, argsR, true, true);
         }
-        _proc.LaunchCommandLineApp(exe, null, wrkDir, false, args, false, true);
+
+        if ((options & (OPT_CUT | OPT_PREPARE | OPT_PREDICT)) != 0)
+        {
+          string exePy = "\"C:/Program Files/Python310/python.exe\"";
+          string datFile = _selectedDir + "/Xdata000.npy";
+          string wrkDir = "C:/Users/chrmu/prj/BatInspector/py";
+          string script = "C:/Users/chrmu/prj/BatInspector/py/batclass.py";
+          string modDir = "C:/Users/chrmu/prj/BatInspector/mod_tsa";
+          string args = script;
+          prepareFolder();
+          if ((options & OPT_CUT) != 0)
+            args += " --cut";
+          if ((options & OPT_PREPARE) != 0)
+            args += " --prepPredict";
+          if ((options & OPT_PREDICT) != 0)
+            args += " --predict";
+          args += " --csvcalls " + reportName +
+               " --root " + modDir + " --specFile " + specFile +
+               " --dataDir " + _selectedDir +
+               " --data " + datFile;
+          retVal = _proc.LaunchCommandLineApp(exePy, null, wrkDir, true, args, true, true);
+        }
       }
+      return retVal;
     }
 
     public List<SpeciesItem> getPossibleSpecies(double charFreq, double duration, double callDist)
@@ -278,6 +317,34 @@ namespace BatInspector
 
       retVal = _proc.IsRunning | _extBusy;
       return retVal;
+    }
+
+    private void createDir(string subDir, bool delete)
+    {
+      string dir = _selectedDir + "/" + subDir;
+      if (!Directory.Exists(dir))
+        Directory.CreateDirectory(dir);
+      if (delete)
+      {
+        System.IO.DirectoryInfo di = new DirectoryInfo(dir);
+
+        foreach (FileInfo file in di.GetFiles())
+        {
+          file.Delete();
+        }
+        foreach (DirectoryInfo d in di.GetDirectories())
+        {
+          d.Delete(true);
+        }
+      }
+    }
+    private void prepareFolder(bool delete = true)
+    {
+      createDir("bat", delete);
+      createDir("wav", delete);
+      createDir("dat", delete);
+      createDir("log", delete);
+      createDir("img", delete);
     }
   }
